refactor(components, protocol-designer): factor out AIR const
Factor the AIR const out of components and into PD. Closes #6110
