+++ protocol-designer/src/components/LiquidsSidebar/index.js
@@ -2,8 +2,9 @@
 import * as React from 'react'
 import { connect } from 'react-redux'
 import { i18n } from '../../localization'
-import { PrimaryButton, SidePanel, swatchColors } from '@opentrons/components'
+import { PrimaryButton, SidePanel } from '@opentrons/components'
 import { PDTitledList } from '../lists'
+import { swatchColors } from '../swatchColors'
 import listButtonStyles from '../listButtons.css'
 
 import { selectors as labwareIngredSelectors } from '../../labware-ingred/selectors'
